Digit/requirements.txt
streamlit
numpy
joblib
pillow
scikit-learn
streamlit-drawable-canvasgit